поменяла тип переменных с int на double, для больших возможностей ввода

+++ src/task/natal/task4/src/Main.java
@@ -3,13 +3,13 @@ public class Main {
     public static void main(String[] args) {
         System.out.println("Введите число х");
         Scanner x = new Scanner(System.in);
-        double num1 = x.nextInt();
+        double num1 = x.nextDouble();
         System.out.println("Введите число у");
         Scanner y = new Scanner(System.in);
-        double num2 = y.nextInt();
+        double num2 = y.nextDouble();
         System.out.println("Введите число z");
         Scanner z = new Scanner(System.in);
-        double num3 = z.nextInt();
+        double num3 = z.nextDouble();
         double a = (num1+num2+num3)/3;
         System.out.println("среднее арифметическое равно " + a);
         int b = (int) (a/3);
